Former-commit-id: d7efb6ff90523143cb8979392bd68f181914779a

+++ Sources/Internal/Scene3D/Systems/QualitySettingsSystem.cpp
@@ -467,7 +467,7 @@ void QualitySettingsSystem::UpdateEntityAfterLoad(Entity *entity)
 	Vector<Entity *> entitiesWithQualityComponent;
 	entity->GetChildEntitiesWithComponent(entitiesWithQualityComponent, Component::QUALITY_SETTINGS_COMPONENT);
 
-    for (uint32 i = 0, sz = entitiesWithQualityComponent.size(); i< sz; ++i)
+    for (size_t i = 0, sz = entitiesWithQualityComponent.size(); i< sz; ++i)
     {
         if (!IsQualityVisible(entitiesWithQualityComponent[i]))
         {
